requirements.txt
instagrapi
google-generativeai
python-dotenv
pillow
